+++ lte/phy/lib/phch/src/sequences.c
@@ -33,7 +33,7 @@
 /**
  * 36.211 6.6.1
  */
-int sequence_pbch(sequence_t *seq, lte_cp_t cp, int cell_id) {
+int sequence_pbch(sequence_t *seq, lte_cp_t cp, uint16_t cell_id) {
   bzero(seq, sizeof(sequence_t));
   return sequence_LTEPRS(seq, CP_ISNORM(cp)?1920:1728, cell_id);
 }
@@ -41,7 +41,7 @@ int sequence_pbch(sequence_t *seq, lte_cp_t cp, int cell_id) {
 /**
  * 36.211 6.7.1
  */
-int sequence_pcfich(sequence_t *seq, int nslot, int cell_id) {
+int sequence_pcfich(sequence_t *seq, uint8_t nslot, uint16_t cell_id) {
   bzero(seq, sizeof(sequence_t));
   return sequence_LTEPRS(seq, 32, (nslot/2+1) * (2*cell_id + 1) * 512 + cell_id);
 }
@@ -50,7 +50,7 @@ int sequence_pcfich(sequence_t *seq, int nslot, int cell_id) {
 /**
  * 36.211 6.9.1
  */
-int sequence_phich(sequence_t *seq, int nslot, int cell_id) {
+int sequence_phich(sequence_t *seq, uint8_t nslot, uint16_t cell_id) {
   bzero(seq, sizeof(sequence_t));
   return sequence_LTEPRS(seq, 12, (nslot/2+1) * (2*cell_id + 1) * 512 + cell_id);
 }
@@ -58,7 +58,7 @@ int sequence_phich(sequence_t *seq, int nslot, int cell_id) {
 /**
  * 36.211 6.8.2
  */
-int sequence_pdcch(sequence_t *seq, int nslot, int cell_id, int len) {
+int sequence_pdcch(sequence_t *seq, uint8_t nslot, uint16_t cell_id, uint32_t len) {
   bzero(seq, sizeof(sequence_t));
   return sequence_LTEPRS(seq, len, (nslot/2) * 512 + cell_id);
 }
@@ -66,7 +66,7 @@ int sequence_pdcch(sequence_t *seq, int nslot, int cell_id, int len) {
 /**
  * 36.211 6.3.1
  */
-int sequence_pdsch(sequence_t *seq, unsigned short rnti, int q, int nslot, int cell_id, int len) {
-	bzero(seq, sizeof(sequence_t));
-	return sequence_LTEPRS(seq, len, (rnti<<14) + (q<<13) + ((nslot/2)<<9) + cell_id);
+int sequence_pdsch(sequence_t *seq, unsigned short rnti, int q, uint8_t nslot, uint16_t cell_id, uint32_t len) {
+  bzero(seq, sizeof(sequence_t));
+  return sequence_LTEPRS(seq, len, (rnti<<14) + (q<<13) + ((nslot/2)<<9) + cell_id);
 }
